+++ src/features/checker/components/ReviewDropdown/ReviewDropdown.stories.tsx
@@ -1,81 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-
-
 import ReviewDropdown from "./ReviewDropdown";
-import { EvaluationSummaryProps } from "./types";
-
-export const mockData: EvaluationSummaryProps[] = [
-  {
-    evaluator: "0x1234567890123456789012345678901234567890",
-    evaluationStatus: "approved",
-    evaluatorType: "human",
-    evaluation: [
-      {
-        question: "How would you rate the quality of the content?",
-        answer: "YES",
-      },
-      {
-        question: "Was the content easy to understand?",
-        answer: "NO",
-      },
-      {
-        question:
-          "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. ",
-        answer: "UNCERTAIN",
-      },
-    ],
-    summary:
-      "The content was well-researched, engaging, and easy to understand. The overall experience was very positive.",
-    lastUpdatedAt: "2024-11-20T13:04:36.042449",
-  },
-  {
-    evaluator: "0x1234567890123456789012345678901234567890",
-    evaluationStatus: "rejected",
-    evaluatorType: "human",
-    evaluation: [
-      {
-        question: "How would you rate the quality of the content?",
-        answer: "YES",
-      },
-      {
-        question: "Was the content easy to understand?",
-        answer: "YES",
-      },
-      {
-        question:
-          "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. ",
-        answer: "UNCERTAIN",
-      },
-    ],
-    summary:
-      "The content was well-researched, engaging, and easy to understand. The overall experience was very positive.",
-    lastUpdatedAt: "2024-11-20T13:04:36.042449",
-  },
-  {
-    evaluator: "0x1234567890123456789012345678901234567890",
-    evaluationStatus: "uncertain",
-    evaluatorType: "llm_gpt3",
-    evaluation: [
-      {
-        question: "How would you rate the quality of the content?",
-        answer: "YES",
-      },
-      {
-        question: "Was the content easy to understand?",
-        answer: "NO",
-      },
-      {
-        question:
-          "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. ",
-        answer: "UNCERTAIN",
-      },
-    ],
-    summary:
-      "The content was well-researched, engaging, and easy to understand. The overall experience was very positive.",
-    lastUpdatedAt: "2024-11-20T13:04:36.042449",
-  },
-];
+import { mockData } from "./mockData";
 
 const meta: Meta<typeof ReviewDropdown> = {
   component: ReviewDropdown,
